fixed bug, changed api url to be able to view all blocks, adjusted effort scaling for effort plot
