--- export
@@ -1,2 +1,0 @@
-export PATH="/usr/local/opt/python@3.9/bin:$PATH"
-
