--- response_df.csv
@@ -1,5 +1,0 @@
-,Time Found,networkDifficulty,variable,value
-0,2024-03-16 06:45:27,421209.62261050055,Response,2.86749
-1,2024-03-19 04:19:20,416310.69022704655,Response,1.8815
-2,2024-03-16 06:45:27,421209.62261050055,effort,2.86749
-3,2024-03-19 04:19:20,416310.69022704655,effort,0.89551
