+++ utils/dash_utils.py
@@ -1,4 +1,3 @@
-from utils.reader import SigmaWalletReader, PriceReader
 from dash import Dash, html, dash_table, dcc
 import dash_bootstrap_components as dbc
